Seperated binning into seperate file
